rework doc tag + minor refactorings
